Fix warnings and jar sealing issues with ITs
* Fix compiler warnings
  * Add missing Deprecated annotation in class method which overrides
    deprecated interface method
  * Add trivial notNull checks in test try-with-resources methods which
    didn't otherwise reference the resource inside the try block
  * Add fail() methods to fail after a test method is called which is
    expected to throw an exception, in order to ensure the exception is
    thrown in the correct line, and to add a reference to the assigned
    variable
  * Add generic parameters and final keyword to CLASS constants in new
    hadoop-mapreduce module
  * Remove import statement for deprecated class
* Rename packages for ITs in new hadoop-mapreduce module to avoid jar
sealing conflicts between the module's main jar and the IT classes
* Fix trivial typos in a comment seen in UserCompactionStrategyIT
